Patch operator deployment
The helm chart doesn't configure emptyDir volumes for `/var/lib/rook`
and `/etc/ceph` on the operator deployment. However, the operator
requires those locations to be writable to provision a Ceph cluster.

This commit adds a post-processing filter which adds emptyDir volumes
for those locations and mounts them in the operator container.