Implement alert rule management
This commit introduces code which takes the upstream Rook.io Ceph
alert rules and creates patched copies of them which work correctly on
OpenShift 4.

Note that this commit relies on the OpenShift 4 alerting behavior that
all alerts without label `syn=true` are silenced automatically.

We rely on the Rook Ceph operator to deploy the required recording rules
and service monitor.