Allow the default SA in the operator namespace to use the Ceph-CSI SCC
This is required for Rook 1.9+ because new DaemonSets which manage the
CSI driver's network namespace have been added in Rook 1.9. These
DaemonSets run with the default ServiceAccount (this is currently not
configurable), so we need to grant the default SA access to the Ceph-CSI
SCC to allow those DaemonSet pods to be created on OpenShift 4.