Enable CSI driver liveness side-car containers
For recent CephCSI versions, the "regular" CSI driver metrics have been
moved to the liveness sidecar. Looking at the metrics, there doesn't
seem to be much of value, so we'll have to look at this in more detail
in a follow-up.

This commit ensures that the configuration produced by the component
doesn't result in Prometheus scrape targets which are down.