Ensure that holder-update DaemonSet pods are restarted on script changes
We add the script's md5sum as an annotation to the DaemonSet pod
template to force Kubernetes to recreate the pods when we modify the
script.

Additionally, we set `sync-wave=10` for the DaemonSet to ensure that the
ConfigMap changes are synced before the checksum annotation on the
DaemonSet is updated.